+++ proto/prysm/v2/keymanager.proto
@@ -1,12 +1,19 @@
 syntax = "proto3";
-package ethereum.validator.accounts.v2;
+package ethereum.prysm.v2;
 
 import "proto/eth/ext/options.proto";
-import "proto/prysm/v1alpha1/attestation.proto";
-import "proto/prysm/v1alpha1/beacon_block.proto";
+import "proto/prysm/v2/attestation.proto";
+import "proto/prysm/v2/beacon_block.proto";
 import "proto/prysm/v2/beacon_block.proto";
 import "google/api/annotations.proto";
 import "google/protobuf/empty.proto";
+
+option csharp_namespace = "Ethereum.Prysm.V2";
+option go_package = "github.com/prysmaticlabs/prysm/proto/prysm/v2;v2";
+option java_multiple_files = true;
+option java_outer_classname = "KeymanagerProto";
+option java_package = "org.ethereum.prysm.v2";
+option php_namespace = "Ethereum\\Prysm\\v2";
 
 // RemoteSigner service API.
 //
@@ -54,10 +61,10 @@ message SignRequest {
     // Beacon chain objects. [100-200]
     oneof object {
         // Phase0 objects.
-        ethereum.eth.v1alpha1.BeaconBlock block = 101;
-        ethereum.eth.v1alpha1.AttestationData attestation_data = 102;
-        ethereum.eth.v1alpha1.AggregateAttestationAndProof aggregate_attestation_and_proof = 103;
-        ethereum.eth.v1alpha1.VoluntaryExit exit = 104;
+        ethereum.prysm.v2.BeaconBlock block = 101;
+        ethereum.prysm.v2.AttestationData attestation_data = 102;
+        ethereum.prysm.v2.AggregateAttestationAndProof aggregate_attestation_and_proof = 103;
+        ethereum.prysm.v2.VoluntaryExit exit = 104;
         uint64 slot = 105 [(ethereum.eth.ext.cast_type) = "github.com/prysmaticlabs/eth2-types.Slot"];
         uint64 epoch = 106 [(ethereum.eth.ext.cast_type) = "github.com/prysmaticlabs/eth2-types.Epoch"];
